requirements.txt
fastapi
uvicorn[standard]
PyJWT
python-dotenv
requests
google-cloud-geminidataanalytics
google-auth
vl-convert-python
